+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Hex color support for `color` and `bgcolor`
+- Hex color support for `color` and `bgcolor` parameters
 - Test system modified
 - Random mode modified
+- `README.md` updated
 ## [0.6] - 2022-04-13
 ### Added
 - `save_params_filter` function