+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `random_hex_color_gen` function
+- `color`,`bgcolor` and `projection` parameters random mode
 ### Changed
 - Hex color support for `color` and `bgcolor` parameters
 - Test system modified
 - Random mode modified
+- `filter_color` function modified
+- `filter_projection` function modified
 - `README.md` updated
 ## [0.6] - 2022-04-13
 ### Added